add prebuilt

+++ scripts/install.sh
@@ -73,26 +73,32 @@ curl --silent 'https://www.amazontrust.com/repository/SFSRootCAG2.pem' --output 
 touch certificate.pem
 touch private.pem.key
 
-if [ ! -d /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c ]
+if [ "$BUILD_KINESIS" == True ]
 then
-  git clone --recursive https://github.com/awslabs/amazon-kinesis-video-streams-webrtc-sdk-c.git
+  if [ ! -d /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c ]
+  then
+    git clone --recursive https://github.com/awslabs/amazon-kinesis-video-streams-webrtc-sdk-c.git
+    cd /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c
+  else
+    cd /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c
+    git pull
+  fi
+
+  git checkout 389703f1f46ce71ab0a77bf9a061feb4a44e9636
+  git submodule update --recursive
+  cd /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c/samples
+  rm /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c/samples/Common.c
+  curl --silent 'https://raw.githubusercontent.com/aws-samples/aws-serverless-telepresence-robot/master/scripts/modified-common.c' --output Common.c
+
   cd /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c
+  mkdir build
+  cd build
+  cmake ..
+  make
+
+  cp /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c/build/kvsWebrtcClientMasterGstSample /home/pi/Projects/robot/
 else
-  cd /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c
-  git pull
+  cd /home/pi/Projects/robot/
+  wget 'https://raw.githubusercontent.com/aws-samples/aws-serverless-telepresence-robot/master/scripts/kvsWebrtcClientMasterGstSample'
+  sudo chmod +x kvsWebrtcClientMasterGstSample
 fi
-
-git checkout 389703f1f46ce71ab0a77bf9a061feb4a44e9636
-git submodule update --recursive
-cd /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c/samples
-rm /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c/samples/Common.c
-curl --silent 'https://raw.githubusercontent.com/aws-samples/aws-serverless-telepresence-robot/master/scripts/modified-common.c' --output Common.c
-
-cd /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c
-mkdir build
-cd build
-cmake ..
-make
-
-cp /home/pi/Projects/amazon-kinesis-video-streams-webrtc-sdk-c/build/kvsWebrtcClientMasterGstSample /home/pi/Projects/robot/
-
